+++ src/main/java/gameObject/Enemy.java
@@ -4,7 +4,7 @@ import interaction.ObjectInteractionVisitor;
 
 public class Enemy extends Ball {
 
-    public Enemy(double _x, double _y, double _speed, double xCoordinate, double yCoordinate, double radius, int key) {
+    public Enemy(double _x, double _y, double _speed, double xCoordinate, double yCoordinate, int radius, int key) {
         super(_x, _y, _speed, xCoordinate, yCoordinate, radius, key);
         type = "E";
     }
