+++ src/main/java/gameObject/Prize.java
@@ -4,7 +4,7 @@ import interaction.ObjectInteractionVisitor;
 
 public class Prize extends Ball {
 
-    public Prize(double _x, double _y, double _speed, double xCoordinate, double yCoordinate, double radius, int key) {
+    public Prize(double _x, double _y, double _speed, double xCoordinate, double yCoordinate, int radius, int key) {
         super(_x, _y, _speed, xCoordinate, yCoordinate, radius, key);
         type = "Pr";
     }
